refactor(web-mobile): split display.jsx into helpers + scoreboard/lobby/streaming
display.jsx (1505 lines) held three display surfaces plus their shared
data/label helpers in one file. Split into five ES modules so streaming, the
per-court scoreboard TV, and the lobby/schedule TV no longer share a file:

- display_helpers.jsx   — shared leaf (findRunningOnCourt, sideLabel,
  queueLabel*, phaseLabel, bracketSidesReady, poolNameOf, TermD, ...)
- display_scoreboard.jsx — per-court TV (TvDisplay + TvWhite/IndividualBoard)
- display_lobby.jsx      — lobby/schedule TV (LobbyDisplay + buildCourtSlots)
- streaming_overlay.jsx  — OBS/vMix overlay (StreamingOverlay, StreamingQR)
- display.jsx            — thin entry: DisplayRoute + barrel re-export of the
  full public surface + the six window.* global assignments

Dependency graph is acyclic (entry -> {scoreboard,lobby,streaming} ->
helpers). Pure move, no behavior change. The barrel re-export keeps all 21
existing exports importable from display.jsx, so the 4 test files and the
window.* global surface are untouched; index.html is unchanged (esbuild
rewrites .jsx->.js and loads the new modules transitively via display.js).

Also remove the worktree web-mobile/node_modules symlink and gitignore it.

Verified: eslint --max-warnings 0, vitest 1505/1505, make go/test (all
packages >=85% coverage), go build.
